requirements.txt
flask>=3.0.0
gunicorn>=21.0.0
mediapipe==0.10.13
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
